requirements.txt
attrs==21.4.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==37.0.1
Cython==0.29.32
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
HTMLParser==0.0.2
ipython==8.18.1
ipython==7.31.1
ipython==8.12.3
ipywidgets==8.1.3
ipywidgets==7.6.5
Jinja2==2.11.3
Jinja2==3.1.3
jnius==1.1.0
keyring==23.4.0
lockfile==0.12.2
matplotlib==3.8.4
matplotlib==3.5.2
mock==4.0.3
networkx==3.2.1
networkx==3.1
networkx==2.8.4
numpy==1.25.2
numpy==1.22.3
numpy==1.24.3
numpy==1.21.5
open3d==0.18.0
pandas==1.4.4
Pillow==9.3.0
Pillow==9.2.0
Pillow==12.0.0
protobuf==4.25.3
protobuf==3.20.3
pyOpenSSL==22.0.0
pyOpenSSL==25.3.0
PyQt5==5.15.11
pyqt5_sip==12.17.1
PyQt5_sip==12.13.0
pywin32==306
pywin32==302
railroad==0.5.0
scikit_learn==1.4.2
scikit_learn==1.0.2
scipy==1.10.1
scipy==1.9.1
Sphinx==5.0.2
thread==2.0.6
torch==2.5.0+cu121
torch==1.9.1
torch==2.3.1
torch==2.2.1
tornado==6.4.1
tornado==6.5.1
tornado==6.1
trove_classifiers==2025.12.1.14
truststore==0.10.4
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
